refactor: renamed package `queue` to `update`
