+++ update/queue.go
@@ -1,4 +1,4 @@
-package queue
+package update
 
 import (
 	"errors"
